# docs/openapi.yml
openapi: 3.0.0
...
info:
  description: Codewind API
  version: 0.1.0
  title: Codewind API
  contact:
    name: API support
    email: example@ibm.com
  license:
    name: Codewind Developer License
    url: 'https://github.com/eclipse/codewind/blob/master/LICENSE'

# tags:
  # - name: users
  #   description: Operations available to regular users
  # - name: admin
  #   description: Operations available only to authenticated admins
paths:
...
  /api/v1/projects/{id}:
    get:
      summary: Get the project object for a single project
      description: Get the project object for the object specified by {id}
      parameters:
        - name: id
          in: path
          schema:
            $ref: '#/components/schemas/ProjectID'
          required: true
          description: id of project
      responses:
        200:
          description: The project was found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        404:
          description: The project was not found
        500:
          description: Any other error
  /api/v1/projects/{id}/loadtest:
    post:
      summary: Run load against a project
      description: Run a load test against a project with the specified id
      parameters:
        - name: id
          in: path
          schema:
            type: string
          required: true
          description: id of project
      requestBody:
        description: An optional description of the test run
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  description: "A description of the test"
                  example: "A new load test"
      responses:
        202:
          description: Request to run-load accepted
        404:
          description: Project not found
        409:
          description: Load test already in progress
        503:
          description: Project is not running
  /api/v1/projects/{id}/loadtest/cancel:
    post:
      summary: Cancels a load-test against a project
      description: Cancel a load-test against a given project
      parameters:
          - name: id
            in: path
            schema:
              type: string
            required: true
            description: id of project
      responses:
        200:
          description: Successful
        404:
          description: Project with given projectID not found
        409:
          description: No run in progress for this project
        500:
          description: Internal error
...
  /api/v1/projects/{id}/loadtest/config:
    post:
      summary: Writes to runLoad/config.json for a project
      description: Writes a valid json to the config file used by loadrunner
      parameters:
        - name: id
          in: path
          schema:
            type: string
          required: true
          description: id of project
      requestBody:
        description: The config.json object, with all fields required
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoadrunnerConfig'
      responses:
        200:
          description: Write to runLoad/config.json successful
          content:
            application/json:
              schema:
                type: string
        400:
          description: Incorrect parameters given
        404:
          description: Project with given projectID not found
        500:
          description: Internal error

    get:
      summary: Reads the runLoad/config.json
      description: Reads the runLoad/config.json. If it does not exist, a file with default parameters will be created
      parameters:
        - name: id
          in: path
          schema:
            type: string
          required: true
          description: id of project
      responses:
        200:
          description: Successful read of config.json file
          content:
            application/json:
              schema:
               $ref: '#/components/schemas/LoadrunnerConfig'
        400:
          description: Incorrect parameters given
        404:
          description: Project not found
        500:
          description: Internal error
  /api/v1/projects/{id}/profiling/{testRunTime}:
    get:
      summary: Reads the runLoad/config.json
      description: Reads the runLoad/config.json. If it does not exist, a file with default parameters will be created
      parameters:
        - name: id
          in: path
          schema:
            type: string
          required: true
          description: id of project
      responses:
        200:
          description: Successful read of config.json file
          content:
            application/json:
              schema:
               $ref: '#/components/schemas/LoadrunnerConfig'
        400:
          description: Incorrect parameters given
        404:
          description: Project not found
        500:
          description: Internal error
...
components:
  schemas:
    Project:
      type: object
      required:
        - projectID
        - name
        - language
        - projectType
        - host
        - ports
      properties:
        projectID:
          $ref: '#/components/schemas/ProjectID'
        name:
          $ref: '#/components/schemas/ProjectName'
        language:
          $ref: '#/components/schemas/ProjectLanguage'
        projectType:
          $ref: '#/components/schemas/ProjectType'
        host:
          type: string
          example: localhost
        ports:
          type: object
          properties:
            exposedPort:
              type: string
              example: "32801"
            internalPort:
              type: string
              example: "3000"
            internalDebugPort:
              type: string
              example: "9229"
    ProjectID:
      type: string
      format: uuid
      example: d290f1ee-6c54-4b01-90e6-d701748f0851
      pattern: '[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}'
    ProjectName:
      description: Name of project, cannot contain any of ", &, /, \\, ?, %, *, :, |, <, > or spaces
      type: string
      example: tobesproject
    ProjectLanguage:
      type: string
    ProjectType:
      description: the build type for the project. Built-in types are 'liberty', 'spring', 'swift', 'nodejs', and 'docker'.
      type: string
...
    LoadrunnerConfig:
      type: object
      required:
        - path
        - requestsPerSecond
        - concurrency
        - maxSeconds
      properties:
        path:
          description: The HTTP endpoint that loadrunner will hit
          type: string
          example: "/loadtest"
        query:
          description: The query parameters to be hit with loadrunner (optional)
          type: object
          example: { a : '1', b : '2' }
        requestsPerSecond:
          description: The rate of requests made by loadrunner
          type: string
          example: "200"
        concurrency:
          description: The number of concurrent threads
          type: string
          example: "200"
        maxSeconds:
          description: The maximum duration for loadrunner to run
          type: string
          example: "20"
        method:
          description: The type of request loadrunner will make
          type: string
          example: "GET"
        body:
          description: A JSON body object that will sent with each load run request
          oneOf:
            - type: object
              example: { id: 1, message: 'codewind' }
            - type: string
              example: "{ id: 1, message: 'codewind'}"
...